requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
openenv-core
openai>=1.14.0
python-dotenv>=1.0.1
torch>=2.0.0
numpy>=1.24.0
